app/requirements.txt
git+https://github.com/CERT-Polska/checkdmarc@allowing-whitespace
dacite==1.8.1
dkimpy==1.1.5
email-validator==2.1.0.post1
fastapi==0.104.1
Jinja2==3.1.2
jinja2-simple-tags==0.5.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-multipart==0.0.6
redis==5.0.1
SQLAlchemy==2.0.23
uvicorn==0.24.0.post1
validators==0.22.0
-r translations/requirements.txt
